+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/club/core/ClubQueryDslRepositoryImpl.java
@@ -10,9 +10,9 @@ import java.util.List;
 
 import static com.mobble.mobbleserver.domain.ClubCategory.QClubCategory.clubCategory;
 import static com.mobble.mobbleserver.domain.address.QAddress.address;
+import static com.mobble.mobbleserver.domain.club.clubGround.QClubGround.clubGround;
 import static com.mobble.mobbleserver.domain.club.core.QClub.club;
 import static com.mobble.mobbleserver.domain.ground.QGround.ground;
-import static com.mobble.mobbleserver.refactor.club.clubGround.entity.QClubGround.clubGround;
 import static com.mobble.mobbleserver.refactor.like.clubLike.entity.QClubLike.clubLike;
 
 @RequiredArgsConstructor
